% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_jupyter.R
\name{cell_to_ipynb}
\alias{cell_to_ipynb}
\title{Converts notebook cells to JSON cells}
\usage{
cell_to_ipynb(cells)
}
\arguments{
\item{cells}{A list of cells.}
}
\value{
A list
}
\description{
Converts notebook cells to JSON cells
}

