% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_jupyter.R
\name{cellImportCheck}
\alias{cellImportCheck}
\title{Checks cell type}
\usage{
cellImportCheck(cell, fileEtx)
}
\arguments{
\item{cell}{Json cell}

\item{fileEtx}{extracted from Json metadata}
}
\value{
character string
}
\description{
Checks cell type
}
