% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_jupyter.R
\name{cellLanguage}
\alias{cellLanguage}
\title{Checks the language of a cell.}
\usage{
cellLanguage(cell)
}
\arguments{
\item{cell}{A single notebook cell}
}
\value{
Language of given cell (string)
}
\description{
Checks the language of a cell.
}
