% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_jupyter.R
\name{cellMagicExt}
\alias{cellMagicExt}
\title{Adds a cell to load extentions required to run R magic cells}
\usage{
cellMagicExt(kernel, cellLanguages, cells)
}
\arguments{
\item{kernel}{kernle}

\item{cellLanguages}{all cell languages in notebook}

\item{cells}{list to append to}
}
\value{
content
}
\description{
Currently only supports R cells in Python, this funtion will be lkely to expand to support new languages
             Appends cell at top of notebook.
}
