% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_jupyter.R
\name{cellToIpynb}
\alias{cellToIpynb}
\title{Converts notebook cells to JSON cells}
\usage{
cellToIpynb(cells)
}
\arguments{
\item{cells}{A list of cells.}
}
\value{
A list
}
\description{
Converts notebook cells to JSON cells
}
