% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_jupyter.R
\name{cellType}
\alias{cellType}
\title{Check if cell is code or markdown}
\usage{
cellType(cell)
}
\arguments{
\item{cell}{A single notebook cell}
}
\value{
cell type
}
\description{
Check if cell is code or markdown
}
