% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_jupyter.R
\name{exportIpynb}
\alias{exportIpynb}
\title{Creates an ipynb list structure and writes to file}
\usage{
exportIpynb(notebook, file = NULL)
}
\arguments{
\item{notebook}{A list.}
}
\value{
A list
}
\description{
Creates an ipynb list structure and writes to file
}
\examples{
notebook <- readRDS("data/notebooks/notebook01.rds")
exportIpynb(notebook, file = "exportedBook.ipynb")
}

