% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_jupyter.R
\name{getKernel}
\alias{getKernel}
\title{Return list in the format used for Jupyter kernel}
\usage{
getKernel(lang = c("R", "Python"))
}
\arguments{
\item{lang}{either 'R' or 'Python'}
}
\value{
list containing either python or R kernel nad language info
}
\description{
Return list in the format used for Jupyter kernel
}
