% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_jupyter.R
\name{importIpynb}
\alias{importIpynb}
\title{Converts Json list to Rcloud notebook}
\usage{
importIpynb(text, filename)
}
\arguments{
\item{text}{Json list}

\item{filename}{filename}
}
\value{
notebook
}
\description{
Converts Json list to Rcloud notebook
}

