% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_jupyter.R
\name{ipyToJson}
\alias{ipyToJson}
\title{Maps Ipynotebook to gist list}
\usage{
ipyToJson(json, filename)
}
\arguments{
\item{json}{Json list}

\item{filename}{filename}
}
\value{
list
}
\description{
Maps Ipynotebook to gist list
}

