% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_jupyter.R
\name{noteBookType}
\alias{noteBookType}
\title{Checks cell type}
\usage{
noteBookType(cell, language)
}
\arguments{
\item{cell}{Json cell}

\item{language}{extracted from Json metadata}
}
\value{
character string
}
\description{
Checks cell type
}
