% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_jupyter.R
\name{shellContent}
\alias{shellContent}
\title{Converts as shell cell to jupyter executable format}
\usage{
shellContent(content)
}
\arguments{
\item{content}{shell cell content}
}
\value{
content
}
\description{
Running shell code in jupyter notebooks is currently only supported in a python kernel
}
