% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_jupyter.R
\name{cell_To_IPYNB}
\alias{cell_To_IPYNB}
\title{Converts notebook cells to JSON cells}
\usage{
cell_To_IPYNB(cells)
}
\arguments{
\item{cells}{A list of cells.}
}
\value{
A list
}
\description{
Converts notebook cells to JSON cells
}
\examples{
notebook <- readRDS("data/notebooks/notebook01.rds")
cell_To_IPYNB(notebook$content$files)
}

