% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_jupyter.R
\name{cellLanguage}
\alias{cellLanguage}
\title{Checks the language of a cell. (First cell supplied)}
\usage{
cellLanguage(cell)
}
\arguments{
\item{cell}{A single notebook cell}
}
\value{
list containing either python or R kernel nad language info
}
\description{
Checks the language of a cell. (First cell supplied)
}
\examples{
notebook <- readRDS("data/notebooks/notebook01.rds")
cellLanguage(notebook$content$files$part1.R)
}

