% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_jupyter.R
\name{cellLanguage}
\alias{cellLanguage}
\title{Checks the language of a cell. (First cell supplied)}
\usage{
cellLanguage(cell, kernel = TRUE)
}
\arguments{
\item{cell}{A single notebook cell}

\item{kernel}{if FALSE just the language name is returned}
}
\value{
list containing either python or R kernel nad language info
}
\description{
Checks the language of a cell. (First cell supplied)
}
